% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{collapse}
\alias{collapse}
\alias{collapse0}
\title{Collapse character Vectors}
\usage{
collapse(x, with = " ")

collapse0(x, with = "")
}
\arguments{
\item{x}{a character vector}

\item{with}{character to place between elements of x.}
}
\description{
Collapse character Vectors
}
