% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comma_list.R
\name{comma_list}
\alias{comma_list}
\title{Construct a comma separated list}
\usage{
comma_list(x, sep = ", ", sep2 = " and ", sep.last = ", and ",
  terminator = "")
}
\arguments{
\item{x}{a list that can be converted into a character.}

\item{sep}{the typical separator}

\item{sep2}{the separator to use in the case of only two elements.}

\item{sep.last}{the separator to use between the last and next to last elements when
there are at least 3 element in the list.}

\item{terminator}{concatenated to the end after the list is concluded.}
}
\description{
Use this utility to create nicely formatted lists for error messages and the like.
}
\examples{
comma_list(c("you", "I"))
comma_list(c("you", "I"), sep2=" & ")
comma_list(head(letters), sep.last=', ', term=', ...')

}
