% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{dot-underscore}
\alias{dot-underscore}
\alias{._}
\title{Alias for gettext/gettextf}
\usage{
._(msg, ..., domain = NULL)
}
\arguments{
\item{msg}{The message to translate.}

\item{...}{Arguments passed on to \code{base::gettextf}
\describe{
  \item{fmt}{a character vector of format strings, each of up to 8192 bytes.}
  \item{domain}{see \code{\link{gettext}}.}
}}

\item{domain}{see [base::gettext]}
}
\description{
This shortcut provides simple translation and formatting functionality.
}
