% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{not-in}
\alias{not-in}
\alias{\%!in\%}
\title{Not in infix operator}
\usage{
x \%!in\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.
    \link{Long vectors} are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against.
    \link{Long vectors} are not supported.}
}
\description{
The same as \code{\link{\%in\%}} but negated.
}
